Update Open Api Specification formatters under Version2 namespace
